1-auto-ba-agent-baseline/requirements.txt
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0
